+++ hardware/bean/avr/cores/bean/Bean.h
@@ -649,7 +649,7 @@ class BeanClass {
    *  Sets the pairing pin for secure modes like HID and ANCS.  The default is 0. 
    *  @param pin the pin to set from 0-65535
    */
-  void setPairingPin(uint16_t pin);
+  void setPairingPin(uint32_t pin);
   ///@}
